fixup! KTOR-8377 Keep compression headers but skip check for relevant platforms
